// src/types.ts
export interface DefaultStyleProps {
    padding?: boolean | number;
    paddingHorizontal?: boolean | number;
    paddingVertical?: boolean | number;
    paddingLeft?: boolean | number;
    paddingRight?: boolean | number;
    paddingTop?: boolean | number;
    paddingBottom?: boolean | number;
    margin?: boolean | number;
    marginHorizontal?: boolean | number;
    marginVertical?: boolean | number;
    marginLeft?: boolean | number;
    marginRight?: boolean | number;
    marginTop?: boolean | number;
    marginBottom?: boolean | number;
}

export interface RnViewProps extends ViewProps, DefaultStyleProps {
    row?: boolean;
    justifyCenter?: boolean;
    justifyStart?: boolean;
    justifyBetween?: boolean,
    justifyEnd?: boolean,
    border?: boolean;
    col?: boolean;
    full?: boolean;
}